pyproject.toml
[project]
name = "test1"
version = "0.1.0"
description = ""
authors = [
    {name = "tigreeroo",email = "ganz67@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
